# src/main/resources/META-INF/openapi.yaml
openapi: 3.0.0
info:
    description: "This is the REST 'contract' for the Order System"
    version: 1.0.0
    title: "(20F) Group Project - Order System"
    contact:
        name: "(20F) CST 8277"
        email: "normanm@algonquincollege.com"
        url: "https://brightspace.algonquincollege.com/d2l/home/279461"
servers:
    -
        url: "http://localhost:{port}/{contextRoot}/{basePath}"
        variables:
            port: {
                default: "8080"
            }
            contextRoot: {
                default: "rest-orderSystem"
            }
            basePath: {
                default: "api/v1"
            }
security:
    -
        basicAuth: []
components:
    securitySchemes:
        basicAuth:
            type: http
            scheme: basic
    parameters:
        idPathParam:
            in: path
            name: id
            required: true
            schema: {
                type: integer,
                format: int64,
                description: "a resource's id",
                example: 1
            }
        orderidPathParam:
            in: path
            name: orderid
            required: true
            schema: {
                type: integer,
                format: int64,
                description: "a resource's id",
                example: 1
            }
        orderlineidPathParam:
            in: path
            name: orderlineid
            required: true
            schema: {
                type: integer,
                format: int64,
                description: "a resource's id",
                example: 1
            }
    responses:
        ForbiddenError:
            description: "Access to the specified resource has been forbidden"
        UnauthorizedAdminError:
            description: "request requires HTTP authentication for ADMIN_ROLE"
        UnauthorizedUserError:
            description: "request requires HTTP authentication for USER_ROLE"
        NotFound:
            description: "The specified resource was not found"
    schemas:
        PojoBase:
            type: object
            required: [id, version]
            properties: {
                id: {
                    description: "id of entity",
                    type: integer
                },
                version: {
                    description: "version (OptLock) of entity",
                    type: integer
                },
                createdDate: {
                    description: "creation timestamp of entity",
                    type: string,
                    format: date-time
                    },
                updatedDate: {
                    description: "last-updatd timestamp of entity",
                    type: string,
                    format: date-time
                }
            }

#Customer
...
        Product:
            description: "(20F) Group Project - Product entity"
            type: object
            allOf: [{
                $ref: "#/components/schemas/PojoBase"
            }]
            properties: {
                description: {type: string, description: "Description", example: Green Apple},
                serialNo: {type: string, description: "serialNo", example: 1156 }
            }
            
#Store            
...
        Order:
            description: "(20F) Group Project - Order entity"
            type: object
            allOf: [{
                $ref: "#/components/schemas/PojoBase"
            }]
            properties: {
                description: {type: string, description: "Description", example: new order}
                
            }

#OrderLine
        OrderLine:
            description: "(20F) Group Project - OrderLine entity"
            type: object
            allOf: [{
                $ref: "#/components/schemas/PojoBase"
            }]
            properties: {
                amount: {type: number, description: "Amount", example: 500.00},
                product: {$ref: '#/components/schemas/Product'}
            }

###################################################################################
#          List of API -- Customer  -- CustomerResource.java -- @Path(customer)   #
###################################################################################

# 
paths:
...
    /customer/{id}/order:
        get:
            summary: "Get all orders by customer id, (ADMIN/USER)"
            parameters: [{
                $ref: "#/components/parameters/idPathParam"
            }]
            responses: {
                "200": {
                    description: "Ok (Successful operation)",
                    content: {
                        application/json: {
                            schema: {
                                $ref: "#/components/schemas/Order"
                            }
                        }
                    }
                },
                "401": {
                    $ref: "#/components/responses/UnauthorizedAdminError"
                }
            }
            
#/customer/{id}/order   Add new order by customer id
# POST
        post:
            summary: "Add new order by customer id, (ADMIN/USER)"
            parameters: [{
                $ref: "#/components/parameters/idPathParam"
            }]
            requestBody: {
                description: "Add new order by its id",
                required: true,
                content: {
                    application/json: {
                        schema: {
                            $ref: "#/components/schemas/Order"
                        }
                    }
                }
            }
            responses: {
                "200": {
                    description: "Ok (Successful operation)",
                    content: {
                        application/json: {
                            schema: {
                                $ref: "#/components/schemas/Order"
                            }
                        }
                    }
                }
            }
            
    # /customer/{id}/order/{orderid}:
    #     get:
    #         summary: "Retrieves a store by its id"
    #         parameters: [{
    #             $ref: "#/components/parameters/idPathParam"
    #         },{
    #             $ref: "#/components/parameters/orderidPathParam"
    #         }]
    #         responses: {
    #             "200": { 
    #                 description: "Ok (Successful operation)",
    #                 content: {
    #                     application/json: {
    #                         schema: {
    #                             $ref: "#/components/schemas/Order"
    #                         }
    #                     }
    #                 }
    #             },
    #             "401": {
    #                 $ref: "#/components/responses/UnauthorizedUserError"
    #             },
    #             "403": {
    #                 $ref: "#/components/responses/ForbiddenError"
    #             },
    #             "404": {
    #                 $ref: "#/components/responses/NotFound"
    #             }
    #         }
        # delete:
        #     summary: "Removes a store by its id"
        #     parameters: [{
        #         $ref: "#/components/parameters/idPathParam"
        #     },{
        #         $ref: "#/components/parameters/orderidPathParam"
        #     }]
        #     responses: {
        #         "204": {
        #             description: "Ok (No content, successful operation)"
        #         },
        #         "401": {
        #             $ref: "#/components/responses/UnauthorizedUserError"
        #         },
        #         "403": {
        #             $ref: "#/components/responses/ForbiddenError"
        #         },
        #         "404": {
        #             $ref: "#/components/responses/NotFound"
        #         }
        #     }
        # put:
        #     summary: "Modified a store by its id"
        #     parameters: [{
        #         $ref: "#/components/parameters/idPathParam"
        #     },{
        #         $ref: "#/components/parameters/orderidPathParam"
        #     }]
        #     requestBody: {
        #         description: "new customer to be modified to Db",
        #         required: true,
        #         content: {
        #             application/json: {
        #                 schema: {
        #                     $ref: "#/components/schemas/Order"
        #                 }
        #             }
        #         }
        #     }
        #     responses: {
        #         "200": {
        #             description: "Ok (Successful operation)",
        #             content: {
        #                 application/json: {
        #                     schema: {
        #                         $ref: "#/components/schemas/Order"
        #                     }
        #                 }
        #             }
        #         },
        #         "401": {
        #             $ref: "#/components/responses/UnauthorizedUserError"
        #         },
        #         "403": {
        #             $ref: "#/components/responses/ForbiddenError"
        #         },
        #         "404": {
        #             $ref: "#/components/responses/NotFound"
        #         }
        #     } 
        
###################################################################################
#          List of API -- Order Line  -- CustomerResource.java                    #
###################################################################################
# @Path(customer)          
#/customer/{id}/order/{orderid}/orderline   Get all order line by order id
#GET
    /customer/{id}/order/{orderid}/orderline:
        get:
            summary: "Retrieves order lines(ADMIN/USER)"
            parameters: [{
                $ref: "#/components/parameters/idPathParam"
            },{
                $ref: "#/components/parameters/orderidPathParam"
            }]
            responses: {
                "200": {
                    description: "Ok (Successful operation)",
                    content: {
                        application/json: {
                            schema: {
                                $ref: "#/components/schemas/OrderLine"
                            }
                        }
                    }
                },
                "401": {
                    $ref: "#/components/responses/UnauthorizedAdminError"
                }
            }
#/customer/{id}/order/{orderid}/orderline   Save order line by order id
#POST
        post:
            summary: "Add order line(ADMIN)"
            parameters: [{
                $ref: "#/components/parameters/idPathParam"
            },{
                $ref: "#/components/parameters/orderidPathParam"
            }]
            requestBody: {
                description: "Add order line",
                required: true,
                content: {
                    application/json: {
                        schema: {
                            $ref: "#/components/schemas/OrderLine"
                        }
                    }
                }
            }
            responses: {
                "200": {
                    description: "Ok (Successful operation)",
                    content: {
                        application/json: {
                            schema: {
                                $ref: "#/components/schemas/OrderLine"
                            }
                        }
                    }
                }
            }
    /customer/{id}/order/{orderid}/orderline/{orderlineid}:
        get:
            summary: "Retrieves order lines(ADMIN/USER)"
            parameters: [{
                $ref: "#/components/parameters/idPathParam"
            },{
                $ref: "#/components/parameters/orderidPathParam"
            },{
                $ref: "#/components/parameters/orderlineidPathParam"
            }]
            responses: {
                "200": {
                    description: "Ok (Successful operation)",
                    content: {
                        application/json: {
                            schema: {
                                $ref: "#/components/schemas/OrderLine"
                            }
                        }
                    }
                },
                "401": {
                    $ref: "#/components/responses/UnauthorizedAdminError"
                }
            }

        put:
            summary: "Modified order lines(ADMIN)"
            parameters: [{
                $ref: "#/components/parameters/idPathParam"
            },{
                $ref: "#/components/parameters/orderidPathParam"
            },{
                $ref: "#/components/parameters/orderlineidPathParam"
            }]
            requestBody: {
                description: "new store to be added to Db",
                required: true,
                content: {
                    application/json: {
                        schema: {
                            $ref: "#/components/schemas/OrderLine"
                        }
                    }
                }
            }
            responses: {
                "200": {
                    description: "Ok (Successful operation)",
                    content: {
                        application/json: {
                            schema: {
                                $ref: "#/components/schemas/OrderLine"
                            }
                        }
                    }
                }
            }
        delete:
            summary: "Delete order lines(ADMIN)"
            parameters: [{
                $ref: "#/components/parameters/idPathParam"
            },{
                $ref: "#/components/parameters/orderidPathParam"
            },{
                $ref: "#/components/parameters/orderlineidPathParam"
            }]
            responses: {
                "204": {
                    description: "Ok (No content, successful operation)"
                },
                "401": {
                    $ref: "#/components/responses/UnauthorizedUserError"
                },
                "403": {
                    $ref: "#/components/responses/ForbiddenError"
                },
                "404": {
                    $ref: "#/components/responses/NotFound"
                }
            }